Feature/search improvement

+++ app/src/main/java/dev/iotarho/artplace/app/ui/favorites/FavoritesFragment.java
@@ -90,8 +90,8 @@ public class FavoritesFragment extends Fragment implements OnFavItemClickListene
 
     @BindView(R.id.fav_artworks_rv)
     RecyclerView favArtworksRv;
-    @BindView(R.id.fav_empty_text)
-    TextView emptyText;
+    @BindView(R.id.empty_screen)
+    View emptyScreen;
     @BindView(R.id.fav_progress_bar)
     ProgressBar favProgressBar;
     @BindView(R.id.fav_coordinator_layout)
@@ -101,6 +101,10 @@ public class FavoritesFragment extends Fragment implements OnFavItemClickListene
     private PagedList<FavoriteArtworks> mFavoriteArtworksList;
     private FavArtworksViewModel mFavArtworksViewModel;
     private String mTitle;
+
+    public static FavoritesFragment newInstance() {
+        return new FavoritesFragment();
+    }
 
     // Required empty public constructor
     public FavoritesFragment() {
@@ -136,7 +140,6 @@ public class FavoritesFragment extends Fragment implements OnFavItemClickListene
         favArtworksRv.setLayoutManager(linearLayoutManager);
 
         mAdapter = new FavArtworkListAdapter(this);
-        emptyText.setVisibility(View.INVISIBLE);
 
         // Show the whole list of Favorite Artworks via the ViewModel
         getFavArtworks();
@@ -165,13 +168,13 @@ public class FavoritesFragment extends Fragment implements OnFavItemClickListene
                 if (favoriteArtworks != null && favoriteArtworks.size() > 0) {
                     // Hide the Progress Bar and the Empty Text View
                     favProgressBar.setVisibility(View.INVISIBLE);
-                    emptyText.setVisibility(View.INVISIBLE);
+                    emptyScreen.setVisibility(View.GONE);
 
                     Log.d(TAG, "Submit artworks to the db " + favoriteArtworks.size());
                     mAdapter.submitList(favoriteArtworks);
                 } else {
                     // Show the empty message when there is no items added to favorites
-                    emptyText.setVisibility(View.VISIBLE);
+                    emptyScreen.setVisibility(View.VISIBLE);
 
                     // Hide the Progress Bar
                     favProgressBar.setVisibility(View.INVISIBLE);
@@ -209,7 +212,7 @@ public class FavoritesFragment extends Fragment implements OnFavItemClickListene
 
     private void refreshFavList() {
 
-        mFavArtworksViewModel.refreshFavArtworkList(Objects.requireNonNull(getActivity()).getApplication()).observe(this, new Observer<PagedList<FavoriteArtworks>>() {
+        mFavArtworksViewModel.refreshFavArtworkList(requireActivity().getApplication()).observe(this, new Observer<PagedList<FavoriteArtworks>>() {
             @Override
             public void onChanged(@Nullable PagedList<FavoriteArtworks> favoriteArtworks) {
                 mAdapter.submitList(favoriteArtworks);
@@ -241,7 +244,7 @@ public class FavoritesFragment extends Fragment implements OnFavItemClickListene
 
         if (mFavoriteArtworksList.size() > 0) {
 
-            AlertDialog.Builder builder = new AlertDialog.Builder(Objects.requireNonNull(getContext()));
+            AlertDialog.Builder builder = new AlertDialog.Builder(requireContext());
             builder.setMessage(R.string.delete_all_message);
             builder.setTitle(R.string.delete_all_title);
             builder.setIcon(R.drawable.ic_delete_24dp);
@@ -280,25 +283,11 @@ public class FavoritesFragment extends Fragment implements OnFavItemClickListene
 
     @Override
     public boolean onOptionsItemSelected(MenuItem item) {
-        switch (item.getItemId()) {
-            case R.id.action_delete_all:
-                // Show warning dialog to the user before deleting all data from the db
-                deleteItemsWithConfirmation();
-                return true;
-
-            default:
-                break;
+        if (item.getItemId() == R.id.action_delete_all) {// Show warning dialog to the user before deleting all data from the db
+            deleteItemsWithConfirmation();
+            return true;
         }
 
         return super.onOptionsItemSelected(item);
     }
-
-    @Override
-    public void onPause() {
-        super.onPause();
-
-        Log.d(TAG, "FavoritesFragment: onPause called");
-    }
-
-
 }
